WEEK5_REPORT/requirements.txt
langchain==0.1.0
openai==1.3.0
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.1.0
numpy==1.24.3
networkx==3.1
pymupdf==1.23.4
pytesseract==0.3.10
Pillow==10.0.0
pyttsx3==2.90
speech_recognition==3.10.0
sounddevice==0.4.6
pydantic-settings==2.1.0
python-dotenv==1.0.0
scipy==1.11.2
